+++ kraftwerk-api/src/main/java/fr/insee/kraftwerk/api/process/AbstractMainProcessingGenesis.java
@@ -43,6 +43,7 @@ public abstract class AbstractMainProcessingGenesis {
     protected UserInputsGenesis userInputs;
     protected final FileUtilsInterface fileUtilsInterface;
     protected Statement database;
+    @Getter
     protected final KraftwerkExecutionContext kraftwerkExecutionContext;
     /* SPECIFIC VARIABLES */
     @Getter
@@ -128,7 +129,7 @@ public abstract class AbstractMainProcessingGenesis {
         int nbPartitions = listIds.size();
         int indexPartition = 1;
         for (List<InterrogationId> listId : listIds) {
-            List<SurveyUnitUpdateLatest> suLatest = client.getUEsLatestState(questionnaireModelId, listId);
+            List<SurveyUnitUpdateLatest> suLatest = client.getResponses(questionnaireModelId, listId);
             log.info("Number of documents retrieved from database : {}, partition {}/{}", suLatest.size(), indexPartition, nbPartitions);
             vtlBindings = new VtlBindings();
             if (dataMode != null){
@@ -142,7 +143,10 @@ public abstract class AbstractMainProcessingGenesis {
     }
 
     protected void unimodalProcess(List<SurveyUnitUpdateLatest> suLatest) throws KraftwerkException {
-        BuildBindingsSequenceGenesis buildBindingsSequenceGenesis = new BuildBindingsSequenceGenesis(fileUtilsInterface);
+        BuildBindingsSequenceGenesis buildBindingsSequenceGenesis = new BuildBindingsSequenceGenesis(
+                fileUtilsInterface,
+                kraftwerkExecutionContext
+        );
         for (String dataMode : userInputs.getModeInputsMap().keySet()) {
             buildBindingsSequenceGenesis.buildVtlBindings(dataMode, vtlBindings, metadataModelsByMode, suLatest, specsDirectory);
             UnimodalSequence unimodal = new UnimodalSequence();