+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/dataprocessing/GroupProcessing.java
@@ -3,6 +3,7 @@ package fr.insee.kraftwerk.core.dataprocessing;
 import fr.insee.bpm.metadata.model.MetadataModel;
 
 import fr.insee.bpm.metadata.model.VariablesMap;
+import fr.insee.kraftwerk.core.utils.KraftwerkExecutionContext;
 import fr.insee.kraftwerk.core.utils.files.FileUtilsInterface;
 import fr.insee.kraftwerk.core.vtl.VtlBindings;
 import fr.insee.kraftwerk.core.vtl.VtlScript;
@@ -11,8 +12,12 @@ public class GroupProcessing extends DataProcessing{
 
 	private MetadataModel metadataModel;
 	
-    public GroupProcessing(VtlBindings vtlBindings, MetadataModel metadataModel, FileUtilsInterface fileUtilsInterface) {
-        super(vtlBindings, fileUtilsInterface);
+    public GroupProcessing(VtlBindings vtlBindings,
+                           MetadataModel metadataModel,
+                           FileUtilsInterface fileUtilsInterface,
+                            KraftwerkExecutionContext kraftwerkExecutionContext
+    ) {
+        super(vtlBindings, fileUtilsInterface, kraftwerkExecutionContext);
         this.metadataModel = metadataModel;
     }