+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/dataprocessing/PaperDataProcessing.java
@@ -3,6 +3,7 @@ package fr.insee.kraftwerk.core.dataprocessing;
 import fr.insee.bpm.metadata.model.MetadataModel;
 import fr.insee.bpm.metadata.model.UcqModality;
 import fr.insee.bpm.metadata.model.UcqVariable;
+import fr.insee.kraftwerk.core.utils.KraftwerkExecutionContext;
 import fr.insee.kraftwerk.core.utils.files.FileUtilsInterface;
 import fr.insee.kraftwerk.core.vtl.VtlBindings;
 import fr.insee.kraftwerk.core.vtl.VtlScript;
@@ -14,8 +15,12 @@ import java.util.List;
  */
 public class PaperDataProcessing extends UnimodalDataProcessing {
 
-    public PaperDataProcessing(VtlBindings vtlBindings, MetadataModel metadataModel, FileUtilsInterface fileUtilsInterface) {
-        super(vtlBindings, metadataModel, fileUtilsInterface);
+    public PaperDataProcessing(VtlBindings vtlBindings,
+                               MetadataModel metadataModel,
+                               FileUtilsInterface fileUtilsInterface,
+                               KraftwerkExecutionContext kraftwerkExecutionContext
+    ) {
+        super(vtlBindings, metadataModel, fileUtilsInterface, kraftwerkExecutionContext);
     }
 
     @Override