+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/dataprocessing/ReconciliationProcessing.java
@@ -1,6 +1,7 @@
 package fr.insee.kraftwerk.core.dataprocessing;
 
 import fr.insee.kraftwerk.core.Constants;
+import fr.insee.kraftwerk.core.utils.KraftwerkExecutionContext;
 import fr.insee.kraftwerk.core.utils.files.FileUtilsInterface;
 import fr.insee.kraftwerk.core.vtl.VtlBindings;
 import fr.insee.kraftwerk.core.vtl.VtlMacros;
@@ -18,14 +19,22 @@ public class ReconciliationProcessing extends DataProcessing {
 	private final String modeVariableIdentifier;
 
 	/** Return processing instance with default mode variable name. */
-	public ReconciliationProcessing(VtlBindings vtlBindings, FileUtilsInterface fileUtilsInterface) {
-		super(vtlBindings, fileUtilsInterface);
+	public ReconciliationProcessing(
+			VtlBindings vtlBindings,
+			FileUtilsInterface fileUtilsInterface,
+			KraftwerkExecutionContext kraftwerkExecutionContext
+	) {
+		super(vtlBindings, fileUtilsInterface, kraftwerkExecutionContext);
 		this.modeVariableIdentifier = Constants.MODE_VARIABLE_NAME;
 	}
 
 	/** Return processing instance with mode variable name given. */
-	public ReconciliationProcessing(VtlBindings vtlBindings, String modeVariableName, FileUtilsInterface fileUtilsInterface) {
-		super(vtlBindings, fileUtilsInterface);
+	public ReconciliationProcessing(VtlBindings vtlBindings,
+									String modeVariableName,
+									FileUtilsInterface fileUtilsInterface,
+									KraftwerkExecutionContext kraftwerkExecutionContext
+	) {
+		super(vtlBindings, fileUtilsInterface, kraftwerkExecutionContext);
 		this.modeVariableIdentifier = modeVariableName;
 	}