+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/dataprocessing/XformsDataProcessing.java
@@ -1,6 +1,7 @@
 package fr.insee.kraftwerk.core.dataprocessing;
 
 import fr.insee.bpm.metadata.model.MetadataModel;
+import fr.insee.kraftwerk.core.utils.KraftwerkExecutionContext;
 import fr.insee.kraftwerk.core.utils.files.FileUtilsInterface;
 import fr.insee.kraftwerk.core.vtl.VtlBindings;
 import fr.insee.kraftwerk.core.vtl.VtlScript;
@@ -10,8 +11,12 @@ import fr.insee.kraftwerk.core.vtl.VtlScript;
  */
 public class XformsDataProcessing extends UnimodalDataProcessing {
 
-    public XformsDataProcessing(VtlBindings vtlBindings, MetadataModel metadataModel, FileUtilsInterface fileUtilsInterface) {
-        super(vtlBindings, metadataModel, fileUtilsInterface);
+    public XformsDataProcessing(VtlBindings vtlBindings,
+                                MetadataModel metadataModel,
+                                FileUtilsInterface fileUtilsInterface,
+                                KraftwerkExecutionContext kraftwerkExecutionContext
+    ) {
+        super(vtlBindings, metadataModel, fileUtilsInterface, kraftwerkExecutionContext);
     }
 
     @Override