+++ kraftwerk-encryption-tests/src/test/java/stubs/GenesisClientStub.java
@@ -67,12 +67,13 @@ public class GenesisClientStub extends GenesisClient {
         }
         return set.stream().toList();
     }
+
     @Override
-    public List<Mode> getModesByQuestionnaire(String questionnaireModelId) throws KraftwerkException {
+    public List<Mode> getModesByQuestionnaire(String questionnaireId) {
         Set<Mode> set = new HashSet<>();
 
         List<SurveyUnitUpdateLatest> filteredMongo = mongoStub.stream().filter(
-                surveyUnitUpdateLatest -> surveyUnitUpdateLatest.getCollectionInstrumentId().equals(questionnaireModelId)
+                surveyUnitUpdateLatest -> surveyUnitUpdateLatest.getCollectionInstrumentId().equals(questionnaireId)
         ).toList();
 
         for (SurveyUnitUpdateLatest surveyUnitUpdateLatest : filteredMongo){
@@ -82,12 +83,12 @@ public class GenesisClientStub extends GenesisClient {
     }
 
     @Override
-    public List<SurveyUnitUpdateLatest> getUEsLatestState(String questionnaireId, List<InterrogationId> interrogationIds) {
+    public List<SurveyUnitUpdateLatest> getResponses(String collectionInstrumentId, List<InterrogationId> interrogationIds) {
         List<SurveyUnitUpdateLatest> list = new ArrayList<>();
 
         List<SurveyUnitUpdateLatest> mongoFiltered1 = mongoStub.stream()
                 .filter(
-                        surveyUnitUpdateLatest -> surveyUnitUpdateLatest.getCollectionInstrumentId().equals(questionnaireId)
+                        surveyUnitUpdateLatest -> surveyUnitUpdateLatest.getCollectionInstrumentId().equals(collectionInstrumentId)
                 ).toList();
 
         Set<SurveyUnitUpdateLatest> mongoFiltered2 = new HashSet<>();
@@ -149,7 +150,6 @@ public class GenesisClientStub extends GenesisClient {
         }
         return entries.getFirst().getValue();
     }
-
     @Override
     @SneakyThrows
     public void saveMetadata(String questionnaireId, Mode mode, MetadataModel metadataModel){
@@ -165,4 +165,5 @@ public class GenesisClientStub extends GenesisClient {
 
         metadataCollectionStub.put(new QuestionnaireIdModeTuple(questionnaireId, mode), metadataModel);
     }
+
 }